Add schemas for verifying and normalizing documentation.
This schema covers all of the plugin documentation.

It has been run on the complete set of Ansible-2.10 docs that
ansible-doc --json is capable of producing at this time.

The schema is intended to convert from invalid input to valid input
whenever the valid input can be extrapolated with a high degree
of accuracy.  In testing, there were only a few collections in
Ansible-2.10 which failed with the converting schema.

We can also generate a json schema from this schema.  Validating
documentation using the json schema will be stricter, failing to
validate whenever something is not as the schema desires.  It's probably
useful to run the json schema on all new plugins to make sure you have
the proper structure for your documentation.  The converting schema can
be used to validate that the old content can be read by the docsite
building tool and can also be used if you want to do something with the
documentation data yourself.

There are two layers for using the schema.  If you import from
schemas.ansible_doc, you have some schemas which are optimized for
working directly with the output from ansible-doc --json.  You can even
pass the json strings you get from that command diectly into it.

If you would like more control (for instance, if you want to publish
documentation from the DOCUMENTATION string even if the RETURN or
EXAMPLES string fails validation), you will want to use the lower level
schemas in the plugin_docs, callback_docs, and module_docs python
modules.