Implement a cli option for specifying an ansible-base cache
Add two new command line options, --ansible-base-cache and
--collections-cache which specify directories where a cached copy of
ansible-base or the collection tarballs live.  If you use these,
antsibull-docs will check whether the cached copy is compatible with and
greater than or equal to the desired version.  If so, it will use the
cached copy instead of downloading its own version.

Fixes #43