Use thread and process worker limits to keep open file count down.
On MacOS, the default open file limit per process is 256.  If we let
antsibull process the full set of collections in parallel we can exceed
this limit.  So we can constrain the number of worker threads so we
dont't hit that limit.

Fixes #44

Future note: once configuration values are added, THREAD_MAX should be
switched from a constant to a config value.