Rename --ansible-base-cache to --ansible-base-source
When content in a cache is out of date, we typically update the copy in
the cache.  When we thought about doing that for the ansible-base
directory passed in by the user, we decided that we should not modify
it as the user is probably passing in their working copy just to save
time on downloading the new version.  That means this is more of an
install source than a cache.  Rename the command line arg appropriately.

Fixes #105