If ansible-base source is passed in, install it rather than running it directly.
This fixes a bug where some pip -e usage doesn't install dependencies of
the package.  This just cropped up recently for me but not for anyone
else.  Unsure of exactly what changed about pip.

This also makes it even harder for us to modify the source accidentally
so it seems good to do regardless of whether the above bug is a general
flaw or only in specific environments.