The .build file needs to record the y-stream info for collections.
In semantic versioning, a version of X.Y.Z has the following meanings:

* X is bumped and Y and Z are set to 0 if there are backwards
  incompatible changes
* Y is bumped and Z is set to 0 if there are new features (and thus
  something that runs with X.Y might not run with X.(Y-1).)
* Z is bumped for bugfixes.  The interface is not changed.

Our policy for new ansible Z-stream releases is that new features may be
added.  So the .builds file was only really recording the information
about X, not Y or Z.

However, our policy for changes to the ansible package between feature
freeze and final (including beta and rc releases) is that we will only
accept bugfixes, not new features, in that timeframe.  So we need to
include the Y version to account for changes in this period.