Add a workflow building nightlies
This change adds a GitHub Actions CI/CD workflow that builds ansible
and ansible-base distribution packages, and then uploads them into a
PyPI-like index at https://ansible-community.github.io/nightly-builds.

The purpose of this effort is to allow third-party projects, that use
ansible and/or ansible-base in any way, to install freshly-cut
bleeding-edge versions of these packages built from the HEADs of their
respective Git repos.

This workflow does not attempt to validate the stability of the
packaged software, meaning that the end-users must be prepared to face
unexpected breaking changes or bugs. There's no expectation that this
software has been thoroughly tested or even tested at all.

The workflow is triggered on every push the main branch and also runs
builds without the publish step on each pull request submitted.

It is also possible to trigger this workflow from the GitHub UI with
the ability to point to different source repositories and their
revisions.

This implementation is not to be considered ideal, nor final and can
be improved over time. Also the decisions regarding the retention
policy, dependency version constraints, and so on are subjects to
change.

Refs:
* https://github.com/ansible/community/issues/539#issuecomment-721964322
* https://github.blog/changelog/2020-07-06-github-actions-manual-triggers-with-workflow_dispatch/
* PR #195

Signed-Off-By: Sviatoslav Sydorenko <webknjaz@redhat.com>