Add logging for the docs macros
We wanted to know how often each of the macros is used in the docs.
Adding logging for that allows us to find out how many macros are used
during a docs run.  We can then parse out the information we need from
the log file.

Here's the script I used with the logs this time:

https://gist.github.com/abadger/aee3e3ef11dd6e8e066feba3484fa946