Revert "Add breadcrumbs to collection index, plugin indexes etc., and add namespace indexes (#240)"
This reverts commit 006ea72bfa003cf43a15e98712d54ba4f1c008fa.

When this commit was applied, docs builds on the jenkins server started
failing.  The error we were presented with wasn't very informative:

08:08:37 Exception occurred:
08:08:37   File "/usr/lib64/python3.6/multiprocessing/connection.py", line 387, in _recv
08:08:37     raise EOFError
08:08:37 EOFError

but we traced it down to the jenkins hosts not having enough memory.
Our working theory is that sphinx uses a lot of memory to process the
toctrees in the breadcrumbs PR.  This causes some of the sphinx workers
to fail when allocating memory.  Because this error is happening in a
subprocess, the main process doesn't see the MemoryError exception, it
only knows that the subprocess quit unexpectedly and isn't returning any
more data on the socket (thus we get EOFError).

Reverting the change for now until we find a way to enable this on the
jenkins nodes.