Remove some commented debugging code
I thought we might turn this into some sort of permanent output that
could cache the state of a docs build at various stages at some later
point but I don't think it fits in anymore.  If we change our mind, the
code is in git so we can always check it out to revive it.