Update the sphinx pygments lexer for sphinx 4
Sphinx 4 drops support for passing an instance of a lexer into
application.add_lexer().  We were only passing an instance so we could
set startinline=True but it looks like the lexer we're basing on doesn't
do anything special with that so it's safe to just use the
AnsibleOutputLexer class instead.