Disable breadcrumbs.
We've run into memory issues on the jenkins builder again.
samccann tested that 7GB is too low but 8GB is enough and will still be
good een if we add 9 more collections with 1.3K+ new plugins.  We're
looking into getting more memory on the jenkins builder but need to
disable breadcrumbs for now, until we get the additional RAM.