build-release: run 'new-ansible' and add more tests (#323)
* build-release: Add a test that validates ansible.in

Collections that are listed in ansible.in should be inside the package.
This could not be the case if "antsibull-build new-ansible" wasn't run
after updating ansible.in.

* build-release: Add a test on the reported ansible version

We are already testing that the version reported by pip is the one we
expect. It is an unexpensive additional sanity check to verify that it's
the same version we are getting from ansible_collections.ansible_version.

* build-release: run 'new-ansible' for alpha and beta versions

Running 'new-ansible' updates the version ranges for every collection in
the 'ansible.in' file. If collections were added to 'ansible.in'
recently, it'll make sure they are added as well.

* build-release: bump default to 5.0.0a3

5.0.0a3 is unreleased which will allow us to test that it works.