Tune down the default threads to 15
Multi-threading improves performance but services we are interacting
with (such as galaxy) are rate limited.

Having a higher number of threads means we are more likely to hammer the
service and get HTTP 429 errors in return which means needing to timeout
and retry which leads to increased overall build time.

Having less threads won't be as fast but should also result in less
HTTP 429s, timeouts and retries so better performance overwall.