build-release: parse and test ansible_builtin_runtime (#330)
* build-release: parse and test ansible_builtin_runtime

This proposes an approach with an Ansible action plugin that is intended
exclusively to parse an ansible_builtin_runtime.yml file. Attempts to do
it in pure Ansible+jinja were made but it felt complicated and fragile.

The plugin retrieves a list of collections and module redirection targets
under plugin_routing.modules which makes it possible to validate whether
or not they are installed with the help of community.general.collection_version
and community.general.a_module.

As a side effect, build-release/tasks/tests.yaml was getting a bit
crowded and so it felt appropriate to split things out into different
files to tidy things up a bit.

Note that for the time being this is meant to run for informational purposes
only and won't fail if there are missing plugins or collections.

* Ignore testns.testcoll.ping in addition to testns.testcoll

* Bump test build version to 5.0.0

* Use latest version of community.general instead of git checkout

The fix we needed for "a_module" was released in 4.0.1, we no longer
need to install from git.