Use poetry-core as the build backend.
poetry has been split into poetry and poetry-core.  The core is a small
subset of functionality that is what is necessary to build python
packages.  Changing from poetry to poetry-core should improve the
`pip install antsibull` experience since all of poetry and its
dependencis won't be needed.

The update to a newer version is also needed to solve:

> https://github.com/python-poetry/poetry/issues/2450

and allow us to put the source in a subdirectory