Move source subdir (#24)
* Use poetry-core as the build backend.

poetry has been split into poetry and poetry-core.  The core is a small
subset of functionality that is what is necessary to build python
packages.  Changing from poetry to poetry-core should improve the
`pip install antsibull` experience since all of poetry and its
dependencis won't be needed.

The update to a newer version is also needed to solve:

> https://github.com/python-poetry/poetry/issues/2450

and allow us to put the source in a subdirectory

* Move source of ansibulled to a subdir.

pyre prefers to be given the directory which holds your python package
as the source-directory rather than the directory which is your python
package.  However, if the directory given is the toplevel, pyre tends to
get slow for me.  (I believe since it scans temporary directories that
I have accumulated [downloads of all those ansible collections...]).
Putting the source code in its own subdirectory is the way to fix that.

* Change the source dir path in github workflows too