Add validate-tags command
This adds a new validate-tags command to retrieve the tags in each
collection's git repository and ensure that the current version is
tagged.

Currently, this is a separate command, but I'd like to have
`antsibull-build` include the data from `get_collections_tags()` in
ansible-build-data and the ansible sdist after this has gotten more
testing/feedback.

I also split CollectionsMetadata into a separate file and added the new
keys. Previously, this class was part of changelog.py and only used for
retrieving changelog URLs.

Relates: https://github.com/ansible-community/community-topics/issues/148
Depends-on: https://github.com/ansible-community/ansible-build-data/pull/176