Add validate-tags command (#456)
* Add validate-tags command

This adds a new validate-tags command to retrieve the tags in each
collection's git repository and ensure that the current version is
tagged.

Currently, this is a separate command, but I'd like to have
`antsibull-build` include the data from `get_collections_tags()` in
ansible-build-data and the ansible sdist after this has gotten more
testing/feedback.

I also split CollectionsMetadata into a separate file and added the new
keys. Previously, this class was part of changelog.py and only used for
retrieving changelog URLs.

Relates: https://github.com/ansible-community/community-topics/issues/148
Depends-on: https://github.com/ansible-community/ansible-build-data/pull/176

* Fix validation of validate-tags options

We should ensure that the input file exists (if that option is passed)
and that the deps file exists to avoid exceptions down the line. Also,
this fixes a typo in _normalize_validate_tags_options.

* Add 456-validate-tags.yaml changelog fragment

* Remove duplicate attribute type defs

* Define tag regex as a file-level constant

* validate_tags: Return list of errors instead of int

* Make get_collections_tags() a library function

Pass needed variables as function arguments instead of using app_ctx.

* Add validate-tags-file subcommand and docstrings

Remove --input from the validate-tags subcommand in favor of a separate
validate-tags-file command.

* get_tags: Set GIT_TERMINAL_PROMPT=0

We don't want git to ask for a password when a repository is
inaccessible. The program should consider it a failure and move on.

* validate-tags: Support a custom tag_version_regex

Some collections don't tag versions as v{version} or {version}. While
collections should use a standard tag name format per the Collection
Requirements, we won't impose this on existing collections.

* Fix type annotation and log collection name