Improve ansible PyPI README
The current README for ansible package on PyPI is not very informative
IMO. Some of it has information that is only relevant to ansible-core.
Additionally, it's missing important details about what the `ansible`
package actually is and how it differs from ansible-core. The Branch
Info section kind of touches on it but it's not detailed and is buried
at the bottom.

- Prominently explain the difference between ansible-core and ansible.
- Remove ansible-core CI badge.
- Add section about reporting issues.
- Mention ansible-build-data

Relates: https://github.com/ansible-community/community-topics/issue/172