Add requirements.txt files for dev dependencies
Previously, the requirements.txt was incongruent with pyproject.toml.
This updates it.

The new requirements.txt makes it easier for developers who wish to
manage venvs manually. This can also be used in CI. Running `pip install
-r requirements.txt` will install antsibull, antsibull-core, and
antsibull-changelog in editable mode, and install all of the test/lint
dependencies. The test and lint dependencies can also be installed
individually.

We can remove the duplicates in pyproject.toml when/if we decide to
ditch Poetry.