validate-tags: allow ignoring certain collections
This adds a `-I` / `--ignore` flag to `antsibull-build validate-tags`
and `antsibull-build validate-tags-file` to ignore errors for specific
collections. `--ignores-file` can be specified separately or
simultaneously to ignore collections from a file with a newline
separated list of names.

Useless ignores are treated as errors by default, but this can be
changed with `--no-warn-useless-ignores`.

`ignores` and `warn_useless_ignores` arguments were added to
`antsibull.tagging.validate_tags()` to support this functionality.