add LegacyBuildMetaMaker; make write-package-files DRY
We still want to build legacy setup.py for ansible 6, 7, and 8. Instead
of duplicating the old code in build_ansible_commands.py, add a
LegacyBuildMetaMaker class with the same interface as BuildMetaMaker.

Note that there's a few differences in the 8.0.0a1 setup.py. I have no
idea why there's extra entries that don't seem to exist in the original
source distribution, but the final wheel's file list is the same.

This change also moves the package file generation code into a function
that can be shared by the rebuild-single and generate-package-files
commands. Typing annotations are changed to StrPath for files where
possible.