build-release: directly install the wheel for testing
We should test the wheel we build and ship to users instead of having
pip build a new one from the sdist. Also, this makes this task more
efficient. Unpacking the sdist, building the wheel, and then unpacking
and installing that wheel takes a while for a large package such as
ansible.