build_meta: remove ansible-core Github repo links
This removes the Bug Tracker and Source Code links from the ansible
package's Python metadata that point to ansible-core repository.

These links are confusing. ansible-core's issue tracker is only for
issues that apply to core, not all of the ansible package. None of the
source code from ansible/ansible is included directly in this package,
either.

The ansible-core links are replaced with a link to the forum and a link
to the ansible-build-data repository.

This only applies to ansible >= 9.0.0rc1.