add announcements subcommand to gen release announcement text  (#573)
* add `announcements` subcommand to gen release announcement text

Co-authored-by: Sviatoslav Sydorenko (Святослав Сидоренко) <webknjaz@redhat.com>

* add basic functional tests for announcements command

* add send-announcements subcommand

Add a new send-announcements process to interactively send announcements
to email, the forum, and Matrix. Instead of requiring release managers
to generate API tokens and input passwords, I chose this compromise
approach. The command automatically opens up pre-filled email and forum
posts and prompts the user about posting to Matrix.

Credits to the Ansible Core developers whose release script that uses
the webbrowser module in a similar way inspired this approach.

* announcements: add --send to automatically send announcements

This makes it easier to chain together the announcements and
send-announcements command.

* use typing_extensions.TypedDict to satisfy pydantic

* announcements: add docstrings

* announcements: make pyperclip dependency optional

* Fix typo.

---------

Co-authored-by: Sviatoslav Sydorenko (Святослав Сидоренко) <webknjaz@redhat.com>
Co-authored-by: Felix Fontein <felix@fontein.de>